ohif-332: extract header component
