+++ platform/ui/src/components/Header/Header.jsx
@@ -1,13 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
+import classNames from 'classnames';
 // TODO: This may fail if package is split from PWA build
 import { useHistory } from 'react-router-dom';
 import { NavBar, Svg, Icon, IconButton, Dropdown } from '@ohif/ui';
 
-function Header({ children, menuOptions }) {
+function Header({ children, menuOptions, isReturnEnabled }) {
   const { t } = useTranslation();
   const history = useHistory();
+
+  const onReturnHandler = () => {
+    if (isReturnEnabled) {
+      history.push('/');
+    }
+  };
 
   return (
     <NavBar className="justify-between border-b-4 border-black">
@@ -16,16 +23,11 @@ function Header({ children, menuOptions }) {
           {/* // TODO: Should preserve filter/sort
               // Either injected service? Or context (like react router's `useLocation`?) */}
           <div
-            className="inline-flex items-center mr-3"
-            onClick={() => history.push('/')}
+            className={classNames("inline-flex items-center mr-3", isReturnEnabled && 'cursor-pointer')}
+            onClick={onReturnHandler}
           >
-            <Icon
-              name="chevron-left"
-              className="w-8 cursor-pointer text-primary-active"
-            />
-            <div className="ml-4 cursor-pointer">
-              <Svg name="logo-ohif" />
-            </div>
+            {isReturnEnabled && <Icon name="chevron-left" className="w-8 text-primary-active" />}
+            <div className="ml-4"><Svg name="logo-ohif" /></div>
           </div>
         </div>
         <div className="flex items-center">{children}</div>
@@ -33,10 +35,7 @@ function Header({ children, menuOptions }) {
           <span className="mr-3 text-lg text-common-light">
             {t('Header:INVESTIGATIONAL USE ONLY')}
           </span>
-          <Dropdown
-            showDropdownIcon={false}
-            list={menuOptions}
-          >
+          <Dropdown showDropdownIcon={false} list={menuOptions}>
             <IconButton
               variant="text"
               color="inherit"
@@ -61,7 +60,12 @@ function Header({ children, menuOptions }) {
 }
 
 Header.propTypes = {
-  children: PropTypes.any.isRequired,
+  children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+  isReturnEnabled: PropTypes.bool
+};
+
+Header.defaultProps = {
+  isReturnEnabled: true
 };
 
 export default Header;
