+++ platform/ui/src/components/Header/index.js
@@ -1,0 +1,2 @@
+import Header from './Header';
+export default Header;
