--- platform/viewer/src/components/PreferencesDropdown/index.js
@@ -1,1 +1,0 @@
-export { default } from './PreferencesDropdown';
